+++ packages/client/src/observable/internal/object/ObjectCacheKey.ts
@@ -25,6 +25,7 @@ import type { ObjectQuery } from "./ObjectQuery.js";
 export const API_NAME_IDX = 0;
 export const PK_IDX = 1;
 export const RDP_CONFIG_IDX = 2;
+export const INCLUDE_ALL_BASE_PROPERTIES_IDX = 3;
 
 export interface ObjectCacheKey extends
   CacheKey<
@@ -35,6 +36,7 @@ export interface ObjectCacheKey extends
       apiName: string,
       pk: PrimaryKeyType<ObjectTypeDefinition>,
       rdpConfig?: Canonical<Rdp> | undefined,
+      includeAllBaseObjectProperties?: true | undefined,
     ]
   >
 {