+++ packages/react/src/public/experimental.ts
@@ -39,7 +39,7 @@ export type {
 } from "../new/useOsdkObjects.js";
 export { useOsdkObjects } from "../new/useOsdkObjects.js";
 export { useRegisterUserAgent } from "../new/useRegisterUserAgent.js";
-export { useOsdkClient } from "../useOsdkClient.js";
+export { useOsdkClient, useOsdkClient2 } from "../useOsdkClient.js";
 export { useOsdkMetadata } from "../useOsdkMetadata.js";
 export type { UseOsdkMetadataResult } from "../useOsdkMetadata.js";
 export { useDebouncedCallback } from "../utils/useDebouncedCallback.js";